@@ -1,6 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
+// import { RouterModule, Routes } from '@angular/router';
+// import { HttpModule } from '@angular/http';
+import {HttpClientModule} from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
@@ -32,7 +34,8 @@ import { StateFlagsService } from './services/state-flags.service';
     imports: [
         AppRoutingModule,
         BrowserModule,
-        HttpModule,
+        // RouterModule,
+        HttpClientModule,
         FormsModule,
         BrowserAnimationsModule,
         PrettyJsonModule,
