@@ -1,20 +1,19 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http } from '@angular/http';
-
-import 'rxjs/add/operator/toPromise';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
+import { Observable, from } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 import { environment } from '../../environments/environment';
 import { AutomationRule } from '../objects/rule-automation';
 import { TimeTrigger } from '../objects/rule-triggers';
 import { ServiceDomain } from '../objects/services';
-import { GrowlService, MessageSeverity } from './growl.service';
 import { StateFlagsService } from './state-flags.service';
-
+import { ListContainer } from "../objects/data-fields";
 
 @Injectable()
 export class OttoRestService {
 
-    private ottoRestUrl = 'http://localhost:5000/rest';
+    private ottoRestUrl: string;
     private headers = new Headers({ 'Content-Type': 'application/json' });
 
     private rules: AutomationRule[] = [];
@@ -25,111 +24,152 @@ export class OttoRestService {
 
 
     constructor(
-        private http: Http,
-        private growl: GrowlService,
+        private http: HttpClient,
         private stateFlags: StateFlagsService,
     ) {
-        let host = environment.ottoHost;
-        let port = environment.ottoPort;
+        const host = environment.ottoHost;
+        const port = environment.ottoPort;
         this.ottoRestUrl = `http://${host}:${port}/rest`;
     }
 
-    getRules(force: boolean = false): Promise<AutomationRule[]> {
+    getRulesObservable(force: boolean = false): Observable<ListContainer<AutomationRule>> {
+        let observable = null;
+        if (force || (this.rules.length === 0)) {
+            // Get a fresh copy of the rules
+            console.log('getRules() fetching from REST API');
 
-        let promise = null;
-
-        if (force || (this.rules.length == 0)) {    // Get a fresh copy of the rules
-            console.log("getRules() fetching from REST API");
-
-            promise = this.http.get(`${this.ottoRestUrl}/rules`)
-                .toPromise()
-                .then(response => {
-                    console.log("getRules() response received");
-                    this.rules = (response.json().data as any[])
-                        .map(rule_obj => {
-                            return AutomationRule.from_object(rule_obj);
-                        });
-                    this.updateRuleIndex(); // Rebuild the rules index
-                    return this.rules.slice(); // Return a copy of the rules
-                })
-                .catch(reason => this.handleRESTError(reason));
+            observable  = this.http.get<OttoRestResponse>(`${this.ottoRestUrl}/rules`)
+                .pipe(
+                    // We have an OttoRestResponse where data: contains an array of rule JSONs
+                    map(resp => resp.data),
+                    // Now we have a single array of rule JSONs
+                    map(rules => rules.map(rule_obj => AutomationRule.from_object(rule_obj))),
+                    // Now we have a single array of AutomationRules
+                    // Wrap it in an AutomationRuleList so it does not get converted into
+                    // a stream of single AutomationRules (consumers expect an array)
+                    map(rules => {
+                        this.rules = rules;
+                        this.updateRuleIndex();
+                        return new ListContainer(rules)
+                    }),
+                );
+        } else {
+            // Return a copy of the cached rules
+            // from() expects a array, so we create a 1 element array with an AutomationRuleList
+            observable = from([new ListContainer(this.rules.slice())]);
         }
-        else {   // Return a copy of the cached rules 
-            promise = Promise.resolve(this.rules.slice());
+        return observable;
+    }
+
+    getRuleObservable(id: string): Observable<AutomationRule> {
+        return this.getRulesObservable()
+            .pipe(
+                // This recieves an AutomationRuleList where .data
+                // has an array of AutomationRule
+                map(rulelist => rulelist.list.filter(rule => rule.id === id)),
+                // Return only the first rule that remains
+                map(rulelist => rulelist[0])
+            );
+    }
+
+    getEntitiesObservable(hidden: boolean = true): Observable<ListContainer<string>> {
+        let observable = null;
+        if (this.entities.length === 0) {
+            // Get a fresh copy of the rules
+            console.log('getEntities() fetching from REST API');
+
+            observable = this.http.get<OttoRestResponse>(`${this.ottoRestUrl}/entities`)
+                .pipe(
+                    map(resp => resp.data),
+                    map(entities => {
+                        if (hidden === false) {
+                            return entities.filter(entity => entity.hidden !== true)
+                        } else {
+                            return entities;
+                        }
+                    }),
+                    map(entities => entities.map(entity => entity['entity_id'])),
+                    map(entities => {
+                        this.entities = entities;
+                        return new ListContainer<string>(entities);
+                    })
+                );
+        } else {
+            // Return a copy of the cached entities
+            observable = from([new ListContainer<string>(this.entities.slice())]);
         }
-
-        return promise;
+        return observable;
     }
 
-    getEntities(hidden: boolean=true): Promise<string[]> {
-        if (this.entities.length == 0) {   // Get a fresh copy of the rules
-            console.log("getEntities() fetching from REST API");
-            return this.http.get(`${this.ottoRestUrl}/entities`)
-                .toPromise()
-                .then(response => {
-                    console.log("getEntities() response received");
-                    this.entities = 
-                        (response.json().data as any[])
-                            .filter(obj => obj.hidden == false)
-                            .map(entry => entry["entity_id"])
-                    return this.entities;
-                })
-                .catch(reason => this.handleRESTError(reason));
+    getServicesObservable(): Observable<ListContainer<ServiceDomain>> {
+        let observable = null;
+        // console.log(this.serviceDomains);
+        if (this.serviceDomains.length === 0) {
+            // Get a fresh copy of the services
+            console.log('getServices() fetching from REST API');
+
+            observable = this.http.get<OttoRestResponse>(`${this.ottoRestUrl}/services`)
+                .pipe(
+                    map(resp => resp.data),
+                    map(data => data.map(obj => ServiceDomain.fromDict(obj))),
+                    map(serviceDomains => {
+                        this.serviceDomains = serviceDomains;
+                        // console.warn(this.serviceDomains);
+                        return new ListContainer<ServiceDomain>(serviceDomains);
+                    })
+                );
+        } else {
+            // Return a copy of the cached services
+            observable = from([new ListContainer<ServiceDomain>(this.serviceDomains)]);
         }
-        // console.log("getEntities serving a cached copy");
-        return Promise.resolve(this.entities.slice());  // Return a copy of the cached entities
+        return observable;
     }
 
-    getServices(): Promise<ServiceDomain[]> {
-        if (this.serviceDomains.length == 0) {
-            console.log("getServices() fetching from REST API");
-            return this.http.get(`${this.ottoRestUrl}/services`)
-                .toPromise()
-                .then(response => {
-                    console.log("getServices() response received");
-                    this.serviceDomains = ServiceDomain.fromRestResponse(response.json().data);
-                    // console.log("serviceDomains.length = " + this.serviceDomains.length);
-                    return this.serviceDomains;
-                })
-                .catch(reason => this.handleRESTError(reason));
+    getZonesObservable(): Observable<string[]> {
+        let observable = null;
+        if (this.entities.length === 0) {
+            // Get a fresh copy of the entities
+            console.log('getZones() fetching from REST API');
+
+            observable = this.getEntitiesObservable()
+                .pipe(
+                    map(entitiesList => entitiesList.list),
+                    map(entities => entities.filter(entity => entity.startsWith('zone.')))
+                );
+        } else {
+            observable = from(this.entities.filter(entity => entity.startsWith('zone.')));
         }
-        return Promise.resolve(this.serviceDomains.slice());
+        return observable;
     }
 
-    getZones(): Promise<string[]> {
-        if (this.entities.length == 0) {   // Get a fresh copy of the entities
-            return this.getEntities()
-                .then((entities) => entities.filter((entity) => entity.startsWith("zone.")));
-        }
-        return Promise.resolve(this.entities.filter((entity) => entity.startsWith("zone.")));
+    getServiceDomainNamesObservable(): Observable<string[]> {
+        return this.getServicesObservable()
+            .pipe(
+                map(domains => domains.list.map(domain => domain.domain))
+            );
     }
 
-    getServiceDomainNames(): Promise<string[]> {
-        return this.getServices().then(domains => this.serviceDomains.map(domain => domain.domain))
+    getServiceNamesObservable(domainName: string): Observable<string[]> {
+        return this.getServicesObservable()
+            .pipe(
+                map(domains => domains.list.filter(
+                    domain => domain.domain === domainName)[0]
+                    .services.map(service => service.service_name)
+                )
+            );
     }
 
-    getServiceNames(domainName: string): Promise<string[]> {
-        return this.getServices().then(domains =>
-            domains.filter(domain => domain.domain == domainName)[0]   // Return only 1 domain
-                .services.map(service => service.service_name)
-        );
-    }
-
-    getRule(id: string): Promise<AutomationRule> {
-        return this.getRules().then(rules => rules.filter(rule => rule.id == id)[0])
-    }
-
-    saveRule(rule: AutomationRule): Promise<OttoRestResponse> {
+    saveRuleObservable(rule: AutomationRule): Observable<OttoRestResponse> {
         this.stateFlags.needsServerReload = true;
 
-        let url = `${this.ottoRestUrl}/rule`;
-        let headers = new Headers();
-        headers.append("Content-Type", "application/json");
+        const url = `${this.ottoRestUrl}/rule`;
+        const headers = new HttpHeaders();
+        headers.append('Content-Type', 'application/json');
 
-        return this.http.put(url, JSON.stringify({ data: rule.toJSON() }), { headers: headers })
-            .toPromise()
-            .then(response => response.json() as OttoRestResponse)
-            .catch(reason => this.handleRESTError(reason));
+        console.log(JSON.stringify({ data: rule.toJSON() }))
+        return this.http.put<OttoRestResponse>(
+            url, JSON.stringify({ data: rule.toJSON() }), { headers: headers }
+        );
     }
 
     addRule(rule: AutomationRule): void {
@@ -140,105 +180,44 @@ export class OttoRestService {
         this.updateRuleIndex();
     }
 
-    deleteRule(ruleId: string): Promise<OttoRestResponse> {
-        let url = `${this.ottoRestUrl}/rule/${ruleId}`;
-        let headers = new Headers();
-        headers.append("Content-Type", "application/json");
+    deleteRuleObservable(ruleId: string): Observable<OttoRestResponse> {
+        const url = `${this.ottoRestUrl}/rule/${ruleId}`;
+        const headers = new HttpHeaders();
+        headers.append('Content-Type', 'application/json');
 
-        return this.http.delete(url, { headers: headers })
-            .toPromise()
-            .then(response => response.json() as OttoRestResponse)
-            .catch(reason => this.handleRESTError(reason));
+        return this.http.delete<OttoRestResponse>(
+            url, { headers: headers });
     }
 
-    enableRule(ruleId: string, enabled: boolean): Promise<OttoRestResponse> {
-        let rule = this.ruleIndex[ruleId] as AutomationRule;
+    enableRuleObservable(ruleId: string, enabled: boolean): Observable<OttoRestResponse> {
+        const rule = this.ruleIndex[ruleId] as AutomationRule;
         rule.enabled = enabled;
-        return this.saveRule(rule);
+        // console.log(`Enabling rule ${ruleId}`);
+        return this.saveRuleObservable(rule);
     }
 
-    serverReloadRules(): Promise<OttoRestResponse> {
-        return this.http.get(`${this.ottoRestUrl}/reload`)
-            .toPromise()
-            .then(response => response.json() as OttoRestResponse)
-            .catch(reason => this.handleRESTError(reason));
+    serverReloadRulesObservable(): Observable<OttoRestResponse> {
+        const url = `${this.ottoRestUrl}/reload`;
+        return this.http.get<OttoRestResponse>(url);
     }
 
+    nextTimeTriggerObservable(trigger: TimeTrigger): Observable<OttoRestResponse> {
+        const url = `${this.ottoRestUrl}/clock/check`;
+        const headers = new HttpHeaders();
+        headers.append('Content-Type', 'application/json');
 
-    updateRuleIndex(): void {
+        return this.http.put<OttoRestResponse>(
+            url, JSON.stringify({ data: trigger.toJSON() }), { headers: headers }
+        );
+    }
+
+    private updateRuleIndex(): void {
         this.ruleIndex = {};
-        for (let rule of this.rules) {
+        for (const rule of this.rules) {
             this.ruleIndex[rule.id] = rule;
         }
     }
-
-    nextTimeTrigger(trigger: TimeTrigger): Promise<OttoRestResponse> {
-        let url = `${this.ottoRestUrl}/clock/check`;
-        let headers = new Headers();
-        headers.append("Content-Type", "application/json");
-
-        return this.http.put(url, JSON.stringify({ data: trigger.toJSON() }), { headers: headers })
-            .toPromise()
-            .then(response => response.json() as OttoRestResponse)
-            .catch(reason => this.handleRESTError(reason));
-    }
-
-    private handleRESTError(reason: any) {
-        let message = 'OttoRESTService: failure during REST request';
-        console.error(message);
-        this.growl.addPersistentMessage(MessageSeverity.ERROR, message, null);
-        return Promise.reject(message);
-        // return null;
-    }
-
-
-
-
-    // getHero(id: number): Promise<Hero> {
-    //   // return this.getHeroes()
-    //   //            .then(heroes => heroes.find(hero => hero.id === id));
-    //   const url = `${this.heroesUrl}/${id}`;
-
-    //   return this.http.get(url)
-    //     .toPromise()
-    //     .then(response => response.json().data as Hero)
-    //     .catch(this.handleError);
-    // }
-
-    // update(hero: Hero): Promise<Hero> {
-    //   const url = `${this.heroesUrl}/${hero.id}`;
-
-    //   return this.http.put(
-    //       url, 
-    //       JSON.stringify(hero), 
-    //       {headers: this.headers}
-    //     )
-    //     .toPromise()
-    //     .then( () => hero)
-    //     .catch(this.handleError);
-    // }
-
-    // create(name: string): Promise<Hero> {
-    //   return this.http
-    //     .post(
-    //       this.heroesUrl,
-    //       JSON.stringify({name: name}),
-    //       {headers: this.headers}
-    //     )
-    //     .toPromise()
-    //     .then(resp => resp.json().data as Hero)
-    //     .catch(this.handleError);
-    // }
-
-    // delete(id: number): Promise<void> {
-    //   const url = `${this.heroesUrl}/${id}`;
-    //   return this.http.delete(url, {headers: this.headers})
-    //     .toPromise()
-    //     .then( () => null)
-    //     .catch(this.handleError);
-    // }
-
-} // class OttoRestService
+}
 
 
 export interface OttoRestResponse {
