@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { Message } from 'primeng/primeng';
+
 import { AutomationRule } from './objects/rule-automation';
 import { GrowlService, MessageSeverity } from './services/growl.service';
 import { OttoRestService } from './services/otto-rest.service';
@@ -11,12 +12,11 @@ import { StateFlagsService } from './services/state-flags.service';
     templateUrl: './app.component.html',
 })
 export class AppComponent {
-    
-    private title = 'Otto UI';
-    // private growlMessages: Message[];
 
     public growl: GrowlService;
     public stateFlags: StateFlagsService;
+
+    private title = 'Otto UI';
 
     constructor(
         private ottoService: OttoRestService,
@@ -25,41 +25,47 @@ export class AppComponent {
         stateFlags: StateFlagsService,
     ) {
         this.growl = growl;
-        this.stateFlags = stateFlags
-        console.log("Constructing app.component");
+        this.stateFlags = stateFlags;
 
-        console.log("Populating OttoRestService's cache");
-        // ottoService.getRules().then(() => 
-        ottoService.getEntities().then(() =>
-            ottoService.getServices().then(() =>
-                console.log("Caches have populated")
-            )
-        )
-        // )
+        ottoService.getEntitiesObservable().subscribe(
+            entities => {
+                console.log('Entities cache populated');
+                ottoService.getServicesObservable().subscribe(
+                    services => {
+                        console.log('Services cache populated');
+                        ottoService.getRulesObservable().subscribe(
+                            rules => console.log('Rules caches have populated')
+                        )
+                    }
+                );
+            }
+        );
+
     }  // constructor
 
-    ngOnInit() {
-        // this.growlMessages = this.growl.messages;
-    }
+    // ngOnInit() {
+    //     // this.growlMessages = this.growl.messages;
+    // }
 
     onAddClick(): void {
-        console.log("Add Rule Clicked");
-        let newRule = new AutomationRule();
+        console.log('Add Rule Clicked');
+        const newRule = new AutomationRule();
         this.ottoService.addRule(newRule);
         this.router.navigate([`/rule/${newRule.id}`]);
     }
 
     onReloadClick(): void {
-        this.ottoService.serverReloadRules()
-            .then(resp => {
-                if (resp.success) {
-                    this.growl.addSuccessMessage(true, resp.message, null);
-                    this.stateFlags.needsServerReload = false;
+        this.ottoService.serverReloadRulesObservable()
+            .subscribe(
+                resp => {
+                    if (resp.success) {
+                        this.growl.addSuccessMessage(true, resp.message, null);
+                        this.stateFlags.needsServerReload = false;
+                    } else {
+                        this.growl.addPersistentMessage(MessageSeverity.ERROR, resp.message, null);
+                    }
                 }
-                else {
-                    this.growl.addPersistentMessage(MessageSeverity.ERROR, resp.message, null);
-                }
-            });
+            );
     }
 
-} // class AppComponent
+}
