@@ -8,13 +8,39 @@ export class RuleActionSequence {
 
   constructor() {}
 
-  // Accessor methods
-  get sequence(): RuleActionItem[] { return this._sequence }
-  get condition(): RuleCondition { return this._condition }
-  get description(): string { return this._description }
-  set description(description: string) {this._description = description }
+  // Static methods
+  static fromObject (obj: any): RuleActionSequence {
+    // console.log("RASeq.fromObject: " + JSON.stringify(obj))
+    const actionSeq: RuleActionSequence = new RuleActionSequence();
 
-  add_action(action:RuleActionItem): void {
+    if ('description' in obj) {
+      actionSeq._description = obj.description;
+    }
+
+    if ('action_condition' in obj) {
+      // console.log("adding action_condition: " + obj.action_condition.condition);
+      actionSeq._condition = RuleCondition.fromObject(obj.action_condition);
+    }
+
+    if ('action_sequence' in obj) {
+      // console.log("adding action_sequence: " + JSON.stringify(obj.action_sequence));
+      for (const actonObj of obj.action_sequence) {
+        // console.log("actionObj: " + JSON.stringify(actonObj));
+        actionSeq.add_action(RuleActionItem.fromObject(actonObj));
+      }
+    }
+    // console.log("Created RuleActionSequence: " + JSON.stringify(actionSeq));
+    return actionSeq;
+  }
+
+
+  // Accessor methods
+  get sequence(): RuleActionItem[] { return this._sequence; }
+  get condition(): RuleCondition { return this._condition; }
+  get description(): string { return this._description; }
+  set description(description: string) {this._description = description; }
+
+  add_action(action: RuleActionItem): void {
     this._sequence.push(action);
   }
 
@@ -40,7 +66,7 @@ export class RuleActionSequence {
 
   // Overridden Property Functions
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.description = this._description;
     if (this._condition != null) {
       o.action_condition = this._condition;
@@ -48,62 +74,24 @@ export class RuleActionSequence {
     o.action_sequence = this._sequence;
     return o;
   }
-
-
-  // Static methods
-  static fromObject (obj: any): RuleActionSequence {
-    // console.log("RASeq.fromObject: " + JSON.stringify(obj))
-    let actionSeq: RuleActionSequence = new RuleActionSequence();
-
-    if ("description" in obj) {
-      actionSeq._description = obj.description;
-    }
-
-    if ("action_condition" in obj) {
-      // console.log("adding action_condition: " + obj.action_condition.condition);
-      actionSeq._condition = RuleCondition.fromObject(obj.action_condition);
-    }
-
-    if ("action_sequence" in obj) {
-      // console.log("adding action_sequence: " + JSON.stringify(obj.action_sequence));
-      for (let actonObj of obj.action_sequence) {
-        // console.log("actionObj: " + JSON.stringify(actonObj));
-        actionSeq.add_action(RuleActionItem.fromObject(actonObj));
-      }
-    }
-
-    // console.log("Created RuleActionSequence: " + JSON.stringify(actionSeq));
-    return actionSeq;
-  }
 }
 
 
 export class RuleActionItem {
 
-  // actionType: string;
-
-  // constructor(actionType: string) {
-  //   this.actionType = actionType;
-  // }
-
   static fromObject(obj: any): RuleActionItem {
-    // console.log("RuleActionItem.fromObject:  " + JSON.stringify(obj));
     let action = null;
 
-    if ("service" in obj) {
+    if ('service' in obj) {
       action = new ServiceAction(obj.domain, obj.service, obj.data);
-    }
-    else if ("condition" in obj) {
+    } else if ('condition' in obj) {
       action = new ConditionAction(RuleCondition.fromObject(obj));
-    }
-    else if ("delay" in obj) {
+    } else if ('delay' in obj) {
       action = new DelayAction(obj.delay);
-    }
-    else if ("log_message" in obj) {
+    } else if ('log_message' in obj) {
       action = new LogAction(obj.log_message);
-    }
-    else {
-      console.error("ERROR: RuleActionItem not matched: " + JSON.stringify(obj));
+    } else {
+      console.error('ERROR: RuleActionItem not matched: ' + JSON.stringify(obj));
     }
 
     return action;
@@ -116,7 +104,7 @@ export class ServiceAction extends RuleActionItem {
   service: string;
   data: any;
 
-  constructor(domain:string, service:string, data:any=null) {
+  constructor(domain: string, service: string, data: any= null) {
     super();
     this.domain = domain;
     this.service = service;
@@ -128,7 +116,7 @@ export class ServiceAction extends RuleActionItem {
 export class ConditionAction extends RuleActionItem {
   condition: RuleCondition;
 
-  constructor(condition:RuleCondition) {
+  constructor(condition: RuleCondition) {
     super();
     this.condition = condition;
   }
@@ -143,7 +131,7 @@ export class ConditionAction extends RuleActionItem {
 export class DelayAction extends RuleActionItem {
   delay: string;
 
-  constructor(delay:string) {
+  constructor(delay: string) {
     super();
     this.delay = delay;
   }
