@@ -11,54 +11,47 @@ export class RuleCondition {
     // if (obj.condition == "always") {
     //   cond = new AlwaysCondition();
     // }
-    if (obj.condition == 'and') {
+    if (obj.condition === 'and') {
       cond = new AndCondition();
-      for (let subcond of obj.conditions) {
+      for (const subcond of obj.conditions) {
         cond.add_condition(RuleCondition.fromObject(subcond));
       }
-    }
-    else if (obj.condition == 'or') {
+    } else if (obj.condition === 'or') {
       cond = new OrCondition();
-      for (let subcond of obj.conditions) {
+      for (const subcond of obj.conditions) {
         cond.add_condition(RuleCondition.fromObject(subcond));
-      }      
-    }
-    else if (obj.condition == 'numeric_state') {
+      }
+    } else if (obj.condition === 'numeric_state') {
       cond = new NumericStateCondition(
         obj.entity_id,
         obj.above_value,
         obj.below_value
-      )
-    }
-    else if (obj.condition == 'state') {
+      );
+    } else if (obj.condition === 'state') {
       cond = new StateCondition(
         obj.entity_id,
         obj.state
-      )
-    }
-    else if (obj.condition == 'sun') {
+      );
+    } else if (obj.condition === 'sun') {
       cond = new SunCondition(
         obj.after,
         obj.before,
         obj.after_offest,
         obj.before_offset
-      )
-    }
-    else if (obj.condition == 'time') {
+      );
+    } else if (obj.condition === 'time') {
       cond = new TimeCondition(
         obj.after,
         obj.before,
         obj.weekday
-      )
-    }
-    else if (obj.condition == 'zone') {
+      );
+    } else if (obj.condition === 'zone') {
       cond = new ZoneCondition(
         obj.entity_id,
         obj.zone
-      )
-    }
-    else { 
-      console.log("ERROR: Rule condition not matched: " + obj.condition);
+      );
+    } else {
+      console.log('ERROR: Rule condition not matched: ' + obj.condition);
     }
     return cond;
   }
@@ -70,8 +63,8 @@ export class ParentCondition extends RuleCondition {
 
   constructor(condition: string) {
     super(condition);
-  }  
-  
+  }
+
   add_condition(condition: RuleCondition): void {
     this.conditions.push(condition);
   }
@@ -97,7 +90,7 @@ export class AndCondition extends ParentCondition {
   conditions: RuleCondition[] = [];
 
   constructor() {
-    super("and");
+    super('and');
   }
 }
 
@@ -106,7 +99,7 @@ export class OrCondition extends ParentCondition {
   conditions: RuleCondition[] = [];
 
   constructor() {
-    super("or"); 
+    super('or');
   }
 }
 
@@ -116,8 +109,8 @@ export class NumericStateCondition extends RuleCondition {
   above_value: number;
   below_value: number;
 
-  constructor(entity_id: string, above_value: number=null, below_value: number=null) {
-    super("numeric_state");
+  constructor(entity_id: string, above_value: number = null, below_value: number = null) {
+    super('numeric_state');
     this.entity_id = entity_id;
     this.above_value = above_value;
     this.below_value = below_value;
@@ -130,17 +123,17 @@ export class StateCondition extends RuleCondition {
   state: string;
 
   constructor(entity_id: string, state: string) {
-    super("state");
+    super('state');
     this.entity_id = entity_id;
     this.state = state;
   }
 
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.condition = this.condition;
     o.entity_id = this.entity_id;
     o.state = this.state;
-    return o;  
+    return o;
   }
 
 }
@@ -152,8 +145,8 @@ export class SunCondition extends RuleCondition {
   after_offset: string;
   before_offset: string;
 
-  constructor(after:string=null, before:string=null, after_offset:string=null, before_offset:string=null) {
-    super("sun");
+  constructor(after: string = null, before: string = null, after_offset: string = null, before_offset: string = null) {
+    super('sun');
     this.after = after;
     this.before = before;
     this.after_offset = after_offset;
@@ -167,8 +160,8 @@ export class TimeCondition extends RuleCondition {
   before: string;
   weekday: string[];
 
-  constructor(after:string=null, before:string=null, weekday:string[]=null){
-    super("time");
+  constructor(after: string = null, before: string = null, weekday: string[ ]= null) {
+    super('time');
     this.after = after;
     this.before = before;
     this.weekday = weekday;
@@ -180,8 +173,8 @@ export class ZoneCondition extends RuleCondition {
   entity_id: string;
   zone: string;
 
-  constructor(entity_id:string, zone:string) {
-    super("zone");
+  constructor(entity_id: string, zone: string) {
+    super('zone');
     this.entity_id = entity_id;
     this.zone = zone;
   }
