@@ -8,37 +8,27 @@ export class RuleTrigger {
     this.platform = platform;
   }
 
-  static fromObject(trig_obj: any): RuleTrigger {    
+  static fromObject(trig_obj: any): RuleTrigger {
     let trig = null;
 
-    // StateTrigger
-    if (trig_obj.platform == 'state') {
+    if (trig_obj.platform === 'state') {
       trig = new StateTrigger(
         trig_obj.entity_id,
         trig_obj.to,
         trig_obj.from
       );
-    }
-
-    // NumericStateTrigger
-    else if (trig_obj.platform == 'numeric_state') {
+    } else if (trig_obj.platform === 'numeric_state') {
       trig = new NumericStateTrigger(
         trig_obj.entity_id,
         trig_obj.above_value,
         trig_obj.below_value
       );
-    }
-
-    // EventTrigger
-    else if (trig_obj.platform == 'event') {
+    } else if (trig_obj.platform === 'event') {
       trig = new EventTrigger(
         trig_obj.event_type,
         trig_obj.event_data
       );
-    }
-
-    // TimeTrigger
-    else if (trig_obj.platform == 'time') {
+    } else if (trig_obj.platform === 'time') {
       trig = new TimeTrigger(
         trig_obj.minute,
         trig_obj.hour,
@@ -47,10 +37,8 @@ export class RuleTrigger {
         trig_obj.weekdays,
         trig_obj.tz
       );
-    }
-
-    else {
-      console.error("ERROR: Trigger platform not matched: " + trig_obj.platform);
+    } else {
+      console.error('ERROR: Trigger platform not matched: ' + trig_obj.platform);
     }
     return trig;
   }
@@ -63,7 +51,7 @@ export class StateTrigger extends RuleTrigger {
   to: string;
   from: string;
 
-  constructor(entity_id: string, to: string=null, from: string=null) {
+  constructor(entity_id: string, to: string = null, from: string = null) {
     super('state');
     this.entity_id = entity_id;
     this.to = to;
@@ -71,7 +59,7 @@ export class StateTrigger extends RuleTrigger {
   }
 
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.platform = this.platform;
     o.entity_id = this.entity_id;
     if (this.to != null && this.to.length > 0) { o.to = this.to; }
@@ -86,7 +74,7 @@ export class NumericStateTrigger extends RuleTrigger {
   above_value: number;
   below_value: number;
 
-  constructor(entity_id: string, above: number=null, below: number=null) {
+  constructor(entity_id: string, above: number = null, below: number = null) {
     super('numeric_state');
     this.entity_id = entity_id;
     this.above_value = above;
@@ -94,11 +82,11 @@ export class NumericStateTrigger extends RuleTrigger {
   }
 
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.platform = this.platform;
     o.entity_id = this.entity_id;
     if (this.above_value != null) { o.above_value = this.above_value; }
-    if (this.below_value != null) { o.below_value = this.below_value; }    
+    if (this.below_value != null) { o.below_value = this.below_value; }
     return o;
   }
 }
@@ -107,7 +95,7 @@ export class NumericStateTrigger extends RuleTrigger {
 export class EventTrigger extends RuleTrigger {
   event_type: string;
   event_data_obj: any;
-  
+
   constructor(event_type: string, event_data_obj: any) {
     super('event');
     this.event_type = event_type;
@@ -115,11 +103,14 @@ export class EventTrigger extends RuleTrigger {
   }
 
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.platform = this.platform;
     o.event_type = this.event_type;
-    if (this.event_data_obj == null) { o.event_data = {}; }
-    else { o.event_data = this.event_data_obj; }
+    if (this.event_data_obj == null) {
+      o.event_data = {};
+    } else {
+      o.event_data = this.event_data_obj;
+    }
     return o;
   }
 }
@@ -132,7 +123,7 @@ export class TimeTrigger extends RuleTrigger {
   weekdays: string;
   tz: string;
 
-  constructor(minute: string, hour: string, day_of_month: string, 
+  constructor(minute: string, hour: string, day_of_month: string,
               month: string, weekdays: string, tz: string) {
     super('time');
     this.minute = minute;
@@ -148,14 +139,14 @@ export class TimeTrigger extends RuleTrigger {
   }
 
   public toJSON = () => {
-    let o: any = {};
+    const o: any = {};
     o.platform = this.platform;
-    if (this.minute != null) { o.minute = this.minute }
-    if (this.hour != null) { o.hour = this.hour }
-    if (this.day_of_month != null) { o.day_of_month = this.day_of_month }
-    if (this.month != null) { o.month = this.month }
-    if (this.weekdays != null) { o.weekdays = this.weekdays }
-    if (this.tz != null) { o.tz = this.tz }
+    if (this.minute != null) { o.minute = this.minute; }
+    if (this.hour != null) { o.hour = this.hour; }
+    if (this.day_of_month != null) { o.day_of_month = this.day_of_month; }
+    if (this.month != null) { o.month = this.month; }
+    if (this.weekdays != null) { o.weekdays = this.weekdays; }
+    if (this.tz != null) { o.tz = this.tz; }
     return o;
   }
 
