@@ -2,11 +2,11 @@ import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { SelectItem } from 'primeng/primeng';
 import {PrettyJsonComponent} from 'angular2-prettyjson';
 
-import { OttoRestService } from '../services/otto-rest.service'
+import { OttoRestService } from '../services/otto-rest.service';
 import { AutomationRule } from '../objects/rule-automation';
-import { 
-  RuleCondition, ParentCondition, AndCondition, OrCondition, StateCondition, 
-  NumericStateCondition, SunCondition, TimeCondition, ZoneCondition 
+import {
+  RuleCondition, ParentCondition, AndCondition, OrCondition, StateCondition,
+  NumericStateCondition, SunCondition, TimeCondition, ZoneCondition
   } from '../objects/rule-conditions';
 import { RuleActionSequence, ConditionAction } from '../objects/rule-actions';
 
@@ -24,8 +24,8 @@ export class RuleConditionComponent implements OnInit {
   @Output() onRemove = new EventEmitter();
 
   debug: boolean = false;
-  longText: string = "xxxxxxxxxxxxxxxx40-charsxxxxxxxxxxxxxxxx";
-  mediumText: string = "xxxxxxx20-charxxxxxx";
+  longText: string = 'xxxxxxxxxxxxxxxx40-charsxxxxxxxxxxxxxxxx';
+  mediumText: string = 'xxxxxxx20-charxxxxxx';
 
   // We need the uiXXX copies of the data in the rule condition
   // because ngModel needs to bind to an object, and if we replace
@@ -37,14 +37,14 @@ export class RuleConditionComponent implements OnInit {
   uiCondition: string;
   uiConditionOptions: SelectItem[];
   uiNestedConditions: RuleCondition[];
-  
+
   // Entity Id
   uiEntityId: string;
   uiEntityIdOptions: SelectItem[];
-  
+
   // State
   uiState: string;
-  
+
   // NumericState
   uiAboveValue: number;
   uiBelowValue: number;
@@ -58,8 +58,8 @@ export class RuleConditionComponent implements OnInit {
 
   // Time
   uiTimeAfter: string;
-  uiTimeBefore: string; 
-  uiTimeWeekday: string[]; 
+  uiTimeBefore: string;
+  uiTimeWeekday: string[];
   uiTimeWeekdayOptions: SelectItem[];
 
   // Zone
@@ -71,7 +71,7 @@ export class RuleConditionComponent implements OnInit {
   ngOnInit(): void {
     // Create Condition options
     this.uiConditionOptions = [];
-    let options = ["and", "or", "state", "numeric_state", "sun", "time", "zone"];
+    let options = ['and', 'or', 'state', 'numeric_state', 'sun', 'time', 'zone'];
     options.map(option => this.uiConditionOptions.push({label: option, value: option}));
     // for (let option of options) {
     //   this.uiConditionOptions.push({label: option, value: option});
@@ -79,7 +79,7 @@ export class RuleConditionComponent implements OnInit {
 
     // Create Sun options
     this.uiSunOptions = [];
-    options = ["sunrise", "sunset"];
+    options = ['sunrise', 'sunset'];
     options.map(option => this.uiSunOptions.push({label: option, value: option}));
     // for (let option of options) {
     //   this.uiSunOptions.push({label: option, value: option});
@@ -87,7 +87,7 @@ export class RuleConditionComponent implements OnInit {
 
     // Create Time Weekday options
     this.uiTimeWeekdayOptions = [];
-    options = ["mon", "tue", "wed", "thu", "fri", "sat", "sun"];
+    options = ['mon', 'tue', 'wed', 'thu', 'fri', 'sat', 'sun'];
     options.map(option => this.uiTimeWeekdayOptions.push({label: option, value: option}));
     // for (let option of options) {
     //   this.uiTimeWeekdayOptions.push({label: option, value: option});
@@ -95,11 +95,13 @@ export class RuleConditionComponent implements OnInit {
 
     // Create Entity ID Options
     this.uiEntityIdOptions = [{label: this.longText, value: null}];
-    this.ottoService.getEntities().then(entities => this.populateOptions(this.uiEntityIdOptions, entities))
+    this.ottoService.getEntitiesObservable()
+      .subscribe(entitiesList => this.populateOptions(this.uiEntityIdOptions, entitiesList.list));
 
     // Create Zone Options
-    this.uiZoneOptions = [{label: this.mediumText, value: null}];    
-    this.ottoService.getZones().then(zones => this.populateOptions(this.uiZoneOptions, zones))
+    this.uiZoneOptions = [{label: this.mediumText, value: null}];
+    this.ottoService.getZonesObservable()
+      .subscribe(zones => this.populateOptions(this.uiZoneOptions, zones));
 
     // Initialize uiXXX components
     if (this.condition != null) {
@@ -109,48 +111,35 @@ export class RuleConditionComponent implements OnInit {
       if (this.condition instanceof StateCondition) {
         this.uiEntityId = this.condition.entity_id;
         this.uiState = this.condition.state;
-      }
-
-      else if (this.condition instanceof NumericStateCondition) {
+      } else if (this.condition instanceof NumericStateCondition) {
         this.uiEntityId = this.condition.entity_id;
         this.uiAboveValue = this.condition.above_value;
         this.uiBelowValue = this.condition.below_value;
-      }
-
-      else if (this.condition instanceof SunCondition) {
+      } else if (this.condition instanceof SunCondition) {
         this.uiSunAfter = this.condition.after;
         this.uiSunAfterOffset = this.condition.after_offset;
         this.uiSunBefore = this.condition.before;
         this.uiSunBeforeOffset = this.condition.before_offset;
-      }
-
-      // Time
-      else if (this.condition instanceof TimeCondition) {
+      } else if (this.condition instanceof TimeCondition) {
         this.uiTimeAfter = this.condition.after;
         this.uiTimeBefore = this.condition.before;
         this.uiTimeWeekday = this.condition.weekday;
-      }
-
-      // Zone
-      else if (this.condition instanceof ZoneCondition) {
+      } else if (this.condition instanceof ZoneCondition) {
         this.uiEntityId = this.condition.entity_id;
         this.uiZone = this.condition.zone;
-      }
-
-      // And & Or
-      else if (this.condition instanceof ParentCondition) {
+      } else if (this.condition instanceof ParentCondition) { // And & Or
         this.uiNestedConditions = this.condition.conditions;
       }
     }
 
-  } // ngOnInit()
+  }
 
 
   populateOptions(option_list: SelectItem[], options: string[]): void {
     option_list.length = 0;   // Clear the array
-    for (let option of options) {
+    for (const option of options) {
       option_list.push({label: option, value: option});
-    }    
+    }
   }
 
 
@@ -158,31 +147,26 @@ export class RuleConditionComponent implements OnInit {
     // When the platform changes, we just re-intitialize the condition
 
     // Set the uiXXX fields to initial values
-    if (this.uiCondition == 'state') {
+    if (this.uiCondition === 'state') {
       this.uiEntityId = null;
       this.uiState = null;
-    }
-    else if (this.uiCondition == 'numeric_state') {
+    } else if (this.uiCondition === 'numeric_state') {
       this.uiEntityId = null;
       this.uiAboveValue = null;
       this.uiBelowValue = null;
-    }
-    else if (this.uiCondition == 'sun') {
+    } else if (this.uiCondition === 'sun') {
       this.uiSunAfter = null;
       this.uiSunAfterOffset = null;
       this.uiSunBefore = null;
       this.uiSunBeforeOffset = null;
-    }
-    else if (this.uiCondition == 'time') {
+    } else if (this.uiCondition === 'time') {
       this.uiTimeAfter = null;
       this.uiTimeBefore = null;
       this.uiTimeWeekday = null;
-    }
-    else if (this.uiCondition == 'zone') {
+    } else if (this.uiCondition === 'zone') {
       this.uiEntityId = null;
       this.uiZone = null;
-    }
-    else if ((this.uiCondition == 'and') || (this.uiCondition == 'or')) {
+    } else if ((this.uiCondition === 'and') || (this.uiCondition === 'or')) {
       this.uiNestedConditions = [];   // initialize to an empty array
     }
     // Now recreate the condition
@@ -197,27 +181,21 @@ export class RuleConditionComponent implements OnInit {
 
   recreateCondition(): void {
     // When the condition changes, we just re-intitialize the condition
-    if (this.uiCondition == 'state') {
+    if (this.uiCondition === 'state') {
       this.replaceCondition(new StateCondition(this.uiEntityId, this.uiState));
-    }
-    else if (this.uiCondition == 'numeric_state') {
+    } else if (this.uiCondition === 'numeric_state') {
       this.replaceCondition(new NumericStateCondition(this.uiEntityId, this.uiAboveValue, this.uiBelowValue));
-    }
-    else if (this.uiCondition == 'sun') {
+    } else if (this.uiCondition === 'sun') {
       this.replaceCondition(new SunCondition(this.uiSunAfter, this.uiSunBefore, this.uiSunAfterOffset, this.uiSunBeforeOffset));
-    }
-    else if (this.uiCondition == 'time') {
+    } else if (this.uiCondition === 'time') {
       this.replaceCondition(new TimeCondition(this.uiTimeAfter, this.uiTimeBefore, this.uiTimeWeekday));
-    }
-    else if (this.uiCondition == 'zone') {
+    } else if (this.uiCondition === 'zone') {
       this.replaceCondition(new ZoneCondition(this.uiEntityId, this.uiZone));
-    }
-    else if (this.uiCondition == 'and') {
+    } else if (this.uiCondition === 'and') {
       this.replaceCondition(new AndCondition());
-    }
-    else if (this.uiCondition == 'or') {
+    } else if (this.uiCondition === 'or') {
       this.replaceCondition(new OrCondition());
-    }    
+    }
   }
 
 
@@ -246,7 +224,7 @@ export class RuleConditionComponent implements OnInit {
     // console.log("Add Condition clicked");
     if (this.condition instanceof ParentCondition) {
       if (this.condition.conditions == null) {
-        this.condition.conditions = []; 
+        this.condition.conditions = [];
       }
       // this.condition.conditions.push(null);
       this.condition.add_condition(null);
@@ -259,7 +237,7 @@ export class RuleConditionComponent implements OnInit {
     //   console.log("ERROR: this.parent is NULL in onRemoveClick");
     //   return;
     // }
-    
+
     // if (this.parent instanceof AutomationRule) {
     //   this.parent.remove_condition();
     // }
@@ -269,7 +247,7 @@ export class RuleConditionComponent implements OnInit {
     // ){
     //   this.parent.conditions.splice(this.parentIndex, 1);
     // }
-    this.onRemove.emit();    
+    this.onRemove.emit();
   }
 
   onConditionReCreate(newCondition: RuleCondition, index: number): void {
@@ -286,4 +264,4 @@ export class RuleConditionComponent implements OnInit {
       }
   }
 
-} // class RuleConditionComponent
+}
